Separated BLL and UI to 2 projects
Instead of a single project, I divided the game to a BLL project and UI
project. That way it will be easier to add additional UIs. (such as WCF
or even an html page if I ever want to put it online)
