+++ .gitignore
@@ -45,5 +45,7 @@ $RECYCLE.BIN/
 Network Trash Folder
 Temporary Items
 .apdisk
-/Othelo/bin
-/Othelo/obj
+/OtheloBLL/bin
+/OtheloBLL/obj
+/OtheloUI/bin
+/OtheloUI/obj
