+++ Othelo.sln
@@ -3,18 +3,43 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
 VisualStudioVersion = 12.0.31101.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Othelo", "Othelo\Othelo.csproj", "{71B10A37-C5B0-4401-B09F-568427E1E042}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OtheloUI", "OtheloUI\OtheloUI.csproj", "{71B10A37-C5B0-4401-B09F-568427E1E042}"
+	ProjectSection(ProjectDependencies) = postProject
+		{F821069D-E374-4222-A553-21EAC2296E71} = {F821069D-E374-4222-A553-21EAC2296E71}
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OtheloBLL", "OtheloBLL\OtheloBLL.csproj", "{F821069D-E374-4222-A553-21EAC2296E71}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|Mixed Platforms = Debug|Mixed Platforms
 		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|Mixed Platforms = Release|Mixed Platforms
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{71B10A37-C5B0-4401-B09F-568427E1E042}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{71B10A37-C5B0-4401-B09F-568427E1E042}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
+		{71B10A37-C5B0-4401-B09F-568427E1E042}.Debug|Mixed Platforms.Build.0 = Debug|x86
 		{71B10A37-C5B0-4401-B09F-568427E1E042}.Debug|x86.ActiveCfg = Debug|x86
 		{71B10A37-C5B0-4401-B09F-568427E1E042}.Debug|x86.Build.0 = Debug|x86
+		{71B10A37-C5B0-4401-B09F-568427E1E042}.Release|Any CPU.ActiveCfg = Release|x86
+		{71B10A37-C5B0-4401-B09F-568427E1E042}.Release|Mixed Platforms.ActiveCfg = Release|x86
+		{71B10A37-C5B0-4401-B09F-568427E1E042}.Release|Mixed Platforms.Build.0 = Release|x86
 		{71B10A37-C5B0-4401-B09F-568427E1E042}.Release|x86.ActiveCfg = Release|x86
 		{71B10A37-C5B0-4401-B09F-568427E1E042}.Release|x86.Build.0 = Release|x86
+		{F821069D-E374-4222-A553-21EAC2296E71}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F821069D-E374-4222-A553-21EAC2296E71}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F821069D-E374-4222-A553-21EAC2296E71}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{F821069D-E374-4222-A553-21EAC2296E71}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{F821069D-E374-4222-A553-21EAC2296E71}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F821069D-E374-4222-A553-21EAC2296E71}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F821069D-E374-4222-A553-21EAC2296E71}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F821069D-E374-4222-A553-21EAC2296E71}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{F821069D-E374-4222-A553-21EAC2296E71}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{F821069D-E374-4222-A553-21EAC2296E71}.Release|x86.ActiveCfg = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
